feat: Introducing pre-defined list for visualisation of tailored results lists

+++ siteroot/controllers/retengine/managers/result_cache.py
@@ -79,11 +79,12 @@ class ResultCache(base_caches.SessionExcludeListCache):
     ## Start of main class implementation
     # ----------------------------------
 
-    def __init__(self, ranklistpath, process_pool, enabled_caches=CacheCfg.all,
+    def __init__(self, predefined_ranklistpath, ranklistpath, process_pool, enabled_caches=CacheCfg.all,
                  enabled_excl_caches=CacheCfg.none):
         """
             Initializes the cache.
             Arguments:
+                predefined_ranklistpath: Path to folder with predefined ranking lists
                 ranklistpath: Path to folder with ranking lists
                 process_pool: pool of workers for multi-threading processing
                 enabled_caches: caches to use in normal operation.
@@ -92,6 +93,8 @@ class ResultCache(base_caches.SessionExcludeListCache):
                                      It should be a valid CacheCfg value.
         """
         self.ranklistpath = ranklistpath
+        self.predefined_ranklistpath = predefined_ranklistpath
+
         # process pool for saving ranking lists in background
         self.process_pool = process_pool
 
@@ -363,33 +366,49 @@ class ResultCache(base_caches.SessionExcludeListCache):
     def _load_results_from_disk(self, query):
         """
             Loads from a local ranking list file the results associated
-            to the specified query.
+            to the specified query. Pre-defined ranking list have priority over
+            ranking list produced by the application.
             Arguments:
                 query: query in dictionary form.
             Returns:
                 The list of results associated to the query, or 'None' if
                 it was not possible to read the file.
         """
+        rlist = None
         fname = self._get_disk_fname(query)
-        if os.path.isfile(fname):
+        predefined_fname = fname.replace(self.ranklistpath, self.predefined_ranklistpath)
+        if os.path.isfile(predefined_fname): # give priority to the predefined list
             try:
-                with open(fname, 'rb') as rfile:
+                with open(predefined_fname, 'rb') as rfile:
                     rlist = msgpack.load(rfile)
-            except:
-                return None
-            return rlist
-        else:
-            return None
+            except Exception as e:
+                rlist = None
+                print e
+                pass
+        if rlist==None:
+            if os.path.isfile(fname):
+                try:
+                    with open(fname, 'rb') as rfile:
+                        rlist = msgpack.load(rfile)
+                except Exception as e:
+                    rlist = None
+                    print e
+                    pass
+        return rlist
 
 
     def _save_results_to_disk(self, query):
         """
-            Saves the results of a query to a local ranking list file.
+            Saves the results of a query to a local ranking list file. If there is
+            already a pre-defined ranking list associated to the query, the results
+            are not saved so as to not overwrite the pre-defined list.
             Arguments:
                 query: query in dictionary form.
         """
         fname = self._get_disk_fname(query)
-        rlist = self._mem_cache.get_results(query)
-
-        with open(fname, 'wb') as rfile:
-            msgpack.dump(rlist, rfile)
+        predefined_fname = fname.replace(self.ranklistpath, self.predefined_ranklistpath)
+        if not os.path.isfile(predefined_fname): # only save it if there is no predefined list
+                                                 # associated to the same query
+            rlist = self._mem_cache.get_results(query)
+            with open(fname, 'wb') as rfile:
+                msgpack.dump(rlist, rfile)
